pyproject.toml
[tool.poetry]
name = "ashley"
version = "0.1.0"
description = ""
authors = ["Carlos Fernandez <carlos.ferlo@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.38.0"
pydantic = "^2.9.2"
llama-index-core = "^0.11.14"
watchdog = ">=2.1.5,<5"
llama-index-llms-cohere = "^0.3.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
